Add custom exception handling to CategoryController

+++ src/main/java/se/iths/crimedatabase/controller/CategoryController.java
@@ -4,6 +4,8 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import se.iths.crimedatabase.entity.Category;
+import se.iths.crimedatabase.exception.BadRequestException;
+import se.iths.crimedatabase.exception.NotFoundException;
 import se.iths.crimedatabase.service.CategoryService;
 
 import java.util.Optional;
@@ -19,18 +21,27 @@ public class CategoryController {
 
     @PostMapping
     public ResponseEntity<Category> create(@RequestBody Category category) {
+        if (category.getName().isEmpty())
+            throw new BadRequestException("Category name cannot be empty");
+
         Category createdCategory = categoryService.create(category);
         return new ResponseEntity<>(createdCategory, HttpStatus.CREATED);
     }
 
     @DeleteMapping("{id}")
     public ResponseEntity<Void> delete(@PathVariable Long id) {
+        if (categoryService.findById(id).isEmpty())
+            throw new NotFoundException(responseMessage(id));
+
         categoryService.delete(id);
         return new ResponseEntity<>(HttpStatus.OK);
     }
 
     @GetMapping("{id}")
     public ResponseEntity<Optional<Category>> findById(@PathVariable Long id) {
+        if (categoryService.findById(id).isEmpty())
+            throw new NotFoundException(responseMessage(id));
+
         Optional<Category> foundCategory = categoryService.findById(id);
         return new ResponseEntity<>(foundCategory, HttpStatus.OK);
     }
@@ -38,6 +49,14 @@ public class CategoryController {
     @GetMapping
     public ResponseEntity<Iterable<Category>> findAll() {
         Iterable<Category> categories = categoryService.findAll();
+        if(!categories.iterator().hasNext())
+            throw new NotFoundException("No categories found");
+
         return new ResponseEntity<>(categories, HttpStatus.OK);
     }
+
+    private String responseMessage(Long id) {
+        return "Category with id: " + id + " cannot be found";
+    }
+
 }
